requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
pydantic
pydantic-settings
torch 
Pillow 
requests
urllib3
python-dotenv
hypothesis>=6.0.0
httpx
pytest
psutil